build_and_test.yml: ensure tests & ./run_clippy.sh pass for build in the Dockerfile
This patch adds a third build type to the build-neon build matrix.
This build type corresponds to what we do in Dockerfile, i.e.,
a `--release` build without `--features testing`.

The advantage of doing it this way is that we can re-use the
steps from the fine-grained & richer build-neon workflow job.

The disadvantage is that we duplicate the build arguments from the
Dockerfile.

The alternative is to do it the other way around, i.e.,
from the Dockerfile, invoke `cargo test` and `./run_clippy.sh`.

However, that would duplicate all the CARGO_FLAGS and CARGO_FEATURES
logic that we have in the build-neon workflow job.
Which seems worse to me.