openapi: document current tenant attach API semantics
We currently return 202 as soon as the tenant is allocated in memory
before we've written out the marker file. So, the /attach API currently
does not have a transactional character. For example, it can happen that
we respond with a 202 and then crash before writing out the marker
file. In such a case, it is important that the client

1. observes the lost attach (by polling tenant status and observing 404)
2. and consequently retries the attach.

It has to do it in this loop until it observes the tenant as "Active"
in the tenant status.  If the client doesn't follow this protocol and instead
goes to another pageserver to attach the tenant, we risk a split-brain situation
where both the first and second pageserver write to the tenant's S3 state.

The improved description highlights the consequences of this
behavior for clients that use the /attach endpoint.

My understanding is that the tenant relocation logic being worked on in
cloud#4740 implements retries, but only until it gets a 202. It will need to
implement the polling after the 202 to be correct.

The motivation for this write-up is that, in a future PR, we'll extend
the /attach endpoint with an option to provide the tenant config.
If we decide to leave the non-transactional behavior of /attach
unmodified, we will be able to avoid persisting the tenant config.
Conversely, if we decide that the /attach API should become
transactional, we'll need to persist the tenant config in the
attach-marker-file before acknowledging receipt of the /attach operation.

refs https://github.com/neondatabase/cloud/pull/4740
refs https://github.com/neondatabase/neon/issues/2238
refs https://github.com/neondatabase/neon/issues/1555