re-fix retryability of timeline_delete
As part of the merge, we reverted the upstream change that
adds a distinguished error type to persist_index_part_with_deleted_flag.
That error type allowed the caller to detect the case where a prior
call to the delete method already succeeded.
That is important to support retrying deletes, as show by the
test case test_delete_timeline_post_rm_failure which failed
as of the merge.

This patch brings back the distinguished error type & handling of it,
thereby getting the test case green again.

Original commit that introduced the distinguished error type:
https://github.com/neondatabase/neon/pull/3919/commits/1312cbc3d943daaf70c56bccb6c7f3f826042b36