replace has_in_progress_downloads with new attachment_status field
Control Plane currently [^1] polls for `has_in_progress_downloads ==
false` after /attach to determine that an attach operation succeeded.

As pointed out in the OpenAPI spec as of neon#4151, polling for
`has_in_progress_downloads` is incorrect.

This patch changes the situation by
- removing `has_in_progress_downloads`
- adding a new field `attachment_status.`
- changing instructions for `/attach` to poll for `attachment_status == attached`.

This makes the instructions in `/attach` actionable for Control Plane.
NB that we don't expose the TenantState in the OpenAPI docs, even though
we expose it in the endpoint. That is with good reason because we
don't want to commit to a fixed set of tenant states forever.
Hence, the separate `attachment_status` field that exposes the bare minimum
required to make /attach + subsequent polling 100% safe wrt split brain.

It would have been nice to report failures explicitly, but the problem
is that we lose that state when we restart. So, we return
`attached` upon attach failure. The tenant is Broken in that case,
causing Control Plane's subsequent health check will fail. Control Plane
can roll back the relocation operation then.
NB: the reliance on the subsequent health check is no change to what we
had before this patch!
NB: we can always add additional TenantAttachmentStatus'es in the future
to communicate failure.

There is one leftover `todo!()` which I don't know how to solve.
See the comments above it for details.
Basically, in the case where we haven't written out a marker file, we
can't transition the tenant into Broken state because that makes us
expose the attachment status `attached`, thereby acknowledging
completion to Control Plane. But, the tenant has no on-disk
representation, so it would be gone on Pageserver restart. To Control Plane,
that situation looks like we have lost the attached tenant. Not great!

[^1]: The one-click tenant relocation PR cloud#4740, currently WIP, is
      the first real user.