make attach marker creation synchronous to the attach handler
Thereby, if we fail to create the marker file, we will not have
created a tenant object / entry in the tenants map.
That saves us from the trouble of figuring out how to remove it
from inside the attach task.

This also adds more transactionality to the attach API.
Not sure if we want to commit to that transactionality, so,
we won't mention it in the OpenAPI docs, and, while in Attaching
state, we will keep attachment_status=maybe even though the
marker file is already written.