refactor: attach: use create_tenant_files + schedule_local_tenant_processing
With this patch, the attach handler now follows the same pattern as
tenant create with regards to instantiation of the new tenant:

1. Prepare on-disk state using `create_tenant_files`.
2. Use the same code path as pageserver startup to load it
   into memory and start background loops
   (`schedule_local_tenant_processing`).

It's a bit sad we can't use the
`PageServerConfig::tenant_attaching_mark_file_path`
method inside `create_tenant_files` because it operates
in a temporary directory. However, it's a small price to pay for the
gained simplicity.

During implementation, I noticed that we don't handle
failures post `create_tenant_files` well. I left TODO
comments in the code linking to the issue that I created
for this [^1].

Also, I'll dedupe the spawn_load and spawn_attach code
in a future commit.

[^1]: https://github.com/neondatabase/neon/issues/4233