Make tenant ID mandatory in tenant_create API call.
We used to generate it, if the caller didn't specify it. That's bad
practice, however, because network is never fully reliable, so it's
possible we create a new tenant but the caller doesn't know about it,
and because it doesn't know the tenant ID, it has no way of retrying
or checking if it succeeded. To discourage that, make it mandatory.
The web control plane has not relied on the auto-generation for a long
time.