don't push latest tags on release branch
before this change we pushed latest tag both on main
and on release. This could lead to akward situations when you
start using some freshly merged to main feature and then the
release happens and you start getting older image. It's possible
to bump into this both locally and in cloud CI.
The proposal is to only push latest on main. This is unlikely to
cause issues because now if someone merges after the release
latest is agin at main. However, the tag is more consistent
and not depend on timing of merges