Create CHANGELOG.md
Add changelog document. The intention is make the Neon Release Notes more focused on the end user. The audience for this changelog will be internal and external developers. Entries in the changelog can be more technical and may even include direct links to PRs.

The initial content that you see in this PR is taken from the Neon Release Notes. Expect the changelog content for future releases to differ as it becomes more technical in nature, written for developers as the primary audience.