to be dropped: tracing-flame second attempt
with configuring it no post processing is needed. the time spent in
loading is very small so one needs to use:

cat .neon/tracing.folded | inferno-flamegraph --minwidth=0 >| flamegraph.svg

To find them, otherwise it will be just walreceiver and compaction.
Noticed also some trouble with the shutdown propagating from a signal
with single threaded runtime; it very likely is that all of the
compactions are already spawned, and as they are mostly blocking code,
they will block the shutdown procedure, which will be quite late
anyways.