test: do immediate stop instead of graceful because upload blocked
good thing is that during the 10s most of the tenants and timelines
would had completed stopping.

one could consider if we should have a timeout here, but the shutdown
freeze_and_flush operations are best-effort. However, if a
timeline_delete is already ongoing, perhaps we should never attempt to
flush it.

this is however already tracked by flush loop being shut down, so there
is no need for more changes.