fix: await detaches and ignores when shutting down pageserver
before this, we just logged the error if there was an ongoing detach or
ignore. however, those are fast, so it's unclear this situation could
ever arise to be noticed.

drop the `_guard` before waiting anything, just in case there's
still a barrier attached to that, and also typo fix.