fix: add an actual eviction error type
this allows us to distinguish different kinds of errors, and not log
huge stacktraces when a file was not found on disk. these have happened
in staging with 1TB databases that compaction::create_image_layers have
kept the layer_removal_cs for a very long time, and deleted a few of the
to-be-evicted layers as part of the compaction.