fix: cancel waiting for semaphore
not too nice implementation, but I didn't want to get into another
argument about cancelation by drop, so let's do it the lot of changes
way.

alternatively would had had to make the localfs safe to drop, which it
is not.

only waits on the semaphore will be cancelled with this trick; forever
uploads can still roam freely.