pageserver: tweak period of imitate_layer_accesses
When the eviction threshold is an integer multiple of the
eviction period, it is unreliable to skip imitating accesses
based on whether the last imitation was more recent than
the threshold.

This is because as finite time passes
between the time used for the periodic execution, and
the 'now' time used for updating last_layer_access_imitation.
When this is just a few milliseconds, and everything else
is on-time, then a 5 second threshold with a 1 second period
will end up entering its 5th iteration slightly _less than_
5 second since last_layer_access_imitation, and thereby
skipping instead of running the imitation.  If a few milliseconds
then pass before we check the access time of a file that
_should_ have been bumped by the imitation pass, then
we end up evicting something we shouldn't have evicted.

This is not a complete solution: it would be good to revisit this
and use a non-walltime mechanism for pinning these layers into
local storage, rather than relying on bumping access times.