fix: remove one needless allocation
I suggested the RawValue usage in the first place, but here it seems to
first create a string and then reqwest creates a Vec of bytes for the
serialization result, so it was essentially not helping except maybe
asserting that we cannot have a serialization issue.

involves adding an unused Clone (+ Copy)? derives to different `Ids`.