pageserver: avoid logging compaction warning during detach
This fixes spurious log lines like:
```
   Compaction failed, retrying in 2s: Cannot run compaction iteration on inactive tenant
```

compaction_loop was doing its first-iteration random_init_delay, but
not checking the CancellationToken between that and calling into
copmaction_iteration, which expects to only be called when the
timeline is in an active state.  We should re-check cancellation
token after sleeping in case we were asked to shut down in the
interim.