pageserver: remove bare tokio::spawn
Commit ddb9c2fe9 added this task, which is launched
with tokio::spawn rather than task_mgr::spawn, and is
not wrapped in a tracing Instrumented<>.

While the task doesn't overtly do any logging, for
consistency we should spawn all our tasks via the wrapper: in
future this task might be extended to e.g. emit log lines that
we would expect to have the proper tracing spans.