pageserver: avoid logging confusing "ERROR" success messages
Example:
```
  walreceiver connection handling ended: db error: ERROR: ending streaming to Some("pageserver") at 0/4031CA8
```

The inner DbError has a severity of ERROR so DbError's Display
implementation includes that ERROR, even though we are actually
logging the error at INFO level.

Introduce an explicit WalReceiverError type, and in its From<>
for postgres errors, apply the logic from ExpectedError, for
expected errors, and a new condition for successes.

The new output looks like:
```
    walreceiver connection handling ended with success: ending streaming to Some("pageserver") at 0/154E9C0, receiver is caughtup and there is no computes
```