tests: shorten the two longest running tests
The overall suite runs comfortably within about 4
minutes when these tests are modified: previously
the runtime is about 10 minutes, most of which is just
spent waiting for these 2 tests to finish.

There's no "right answer" for how long a test should
be, but as a rule of thumb, no one test should run
for much longer than the time it takes to run all the
other tests together.