warn instead of panic on incorrect option config
warn when --vm-monitor-addr is set but AUTOSCALING is not instead of
panic'ing. This is currently caused because the monitor arguments have
default arguments. When the arguments are set by the control plane,
we can remove the defaults. We can also change it back to a panic if
that is determined to be the correct course.