gracefully shutdown monitor with CancellationToken
All monitor tokio::spawn'ed threads now listen for a cancel signal and
will gracefully shutdown upon receiving it. This is not technically
necessary as the VM get's completely destroyed when the compute is
shut down. This is probably safer w.r.t. resources such as file
handles and network connections.