fix: handle remote only future layers
confusing situation with test_remote_storage_upload_queue_retries
documented on the PR discussion. an image layer was remote only and it
took me again a while to recollect how is this possible to have future
image layers (it is because we can create them for not yet flushed to
disk lsns).